+++ app/repositories/Repository.php
@@ -2,12 +2,13 @@
 
 namespace App\Repositories;
 
+use App\Core\App;
 use App\Core\Database;
 
 class Repository
 {
     protected static function db()
     {
-        return new Database();
+        return App::resolve(Database::class);
     }
 }